requirements.txt
youtube-transcript-api
google-generativeai
Flask
dotenv